services/backfill-worker-service/requirements.txt
pika==1.3.2
yfinance>=0.2.48
python-dotenv==1.0.1
pandas>=1.5.0
requests>=2.28.0
scipy>=1.9.0
psycopg2-binary==2.9.9
alpaca-trade-api
alpha_vantage==2.3.1
alpaca-py
